Refactor loading synonyms from index into a separate function
