+++ server/src/main/java/org/elasticsearch/index/analysis/Analysis.java
@@ -43,15 +43,20 @@ import org.apache.lucene.analysis.ru.RussianAnalyzer;
 import org.apache.lucene.analysis.sv.SwedishAnalyzer;
 import org.apache.lucene.analysis.th.ThaiAnalyzer;
 import org.apache.lucene.analysis.tr.TurkishAnalyzer;
+import org.elasticsearch.action.support.PlainActionFuture;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.logging.DeprecationCategory;
 import org.elasticsearch.common.logging.DeprecationLogger;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
+import org.elasticsearch.synonyms.SynonymRule;
+import org.elasticsearch.synonyms.SynonymsManagementAPIService;
+import org.elasticsearch.threadpool.ThreadPool;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.Reader;
+import java.io.StringReader;
 import java.nio.charset.CharacterCodingException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -290,6 +295,32 @@ public class Analysis {
             String message = String.format(Locale.ROOT, "IOException while reading %s_path: %s", settingPrefix, path.toString());
             throw new IllegalArgumentException(message, ioe);
         }
+    }
+
+    public static Reader getReaderFromIndex(
+        String synonymsSet,
+        ThreadPool threadPool,
+        SynonymsManagementAPIService synonymsManagementAPIService
+    ) {
+        // TODO: this is a temporary solution for loading synonyms under feature flag, to be redesigned for GA
+        final PlainActionFuture<SynonymsManagementAPIService.SynonymsSetResult> synonymsLoadingFuture = new PlainActionFuture<>() {
+            @Override
+            protected boolean blockingAllowed() {
+                // allow blocking while loading synonyms under feature flag
+                return true;
+            }
+        };
+        threadPool.executor(ThreadPool.Names.SYSTEM_READ).execute(() -> {
+            synonymsManagementAPIService.getSynonymsSet(synonymsSet, 0, 10_000, synonymsLoadingFuture);
+        });
+        final SynonymsManagementAPIService.SynonymsSetResult results = synonymsLoadingFuture.actionGet();
+
+        SynonymRule[] synonymRules = results.synonymRules();
+        StringBuilder sb = new StringBuilder();
+        for (SynonymRule synonymRule : synonymRules) {
+            sb.append(synonymRule.synonyms()).append(System.lineSeparator());
+        }
+        return new StringReader(sb.toString());
     }
 
 }
